requirements.txt
streamlit
pandas
numpy
scikit-learn
jobli